test: reorganize tests to match new tool module structure
- Create test_table.py from test_introspection.py (table tests) + test_utility.py
- Create test_record.py from test_introspection.py (record_get) + test_relationships.py
- Create test_record_write.py from test_developer.py
- Delete test_dev_utils.py (tools removed)
- Update test_server.py: introspection_only -> core_readonly
- Update test_packages.py: new group names and counts
- 1034 tests passing
